DRUM_BRAIN 1.0 Update - use webmidi
this webpage can be used as a source for the webmidi browser to feed
improvised drum patterns into midi drum apps. The top row of
checkboxes on each drum is the “always hit” row and will play every
time thru your loop. the bottom row is the “maybe hit” row. The lower
row of checkboxes will play some times and skip other times. the
result is a drum machine that never needs to repeat itself that can be
used with any midi receiving device or app. example: use the MFmonitor
as the output of the drum_brain webpage in the Webmidi Browser on IOS,
and then allow the DM1 drum app to receive midi notes in from
MFmonitor app.